src/ghosts.pandora/requirements.txt
fastapi[standard]>=0.115.0,<0.116.0
uvicorn==0.32.0
pydantic==2.9.2
Faker==30.6.0
fpdf==1.7.2
opencv-python==4.10.0.84
openpyxl==3.1.5
pyzipper==0.3.6
zipstream==1.1.4
python-docx==1.1.2
python-pptx==1.0.2
rarfile==4.2
requests==2.32.3
diffusers==0.31.0
transformers==4.46.3
pillow==11.0.0
accelerate==1.1.1
scipy==1.14.1
safetensors==0.4.5
numpy==2.1.3
torch
pyttsx3==2.98
protobuf
sentencepiece==0.2.0
imaginairy
python-json-logger
torchao
pytest
pytest-mock
bs4
py_mini_racer
cssutils
